pyproject.toml
[tool.poetry]
name = "cookify"
version = "0.1.0"
description = ""
authors = ["Luca Dalla Valle <luca.dallavalle@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.7.0"}
python-magic = "^0.4.27"


[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
